address PR comments:
* move check for UINT24_MAX to reserve write 24
* define UINT24_MAX in those two places where they are needed
* define UINT24_MAX with 1 << 24
* move the test from handshake test to handshake misc test
* check for < not <= in assertion
