+++ tests/unit/s2n_handshake_test.c
@@ -30,7 +30,6 @@
 #include "tls/s2n_cipher_suites.h"
 #include "tls/s2n_connection.h"
 #include "tls/s2n_security_policies.h"
-#include "tls/s2n_tls.h"
 #include "tls/s2n_tls13.h"
 #include "utils/s2n_safety.h"
 
@@ -233,31 +232,10 @@ int async_pkey_fn(struct s2n_connection *conn, struct s2n_async_pkey_op *op)
     return 0;
 }
 
-static S2N_RESULT s2n_test_handshake_finish_header(uint32_t handshake_message_length)
-{
-    DEFER_CLEANUP(struct s2n_stuffer stuffer, s2n_stuffer_free);
-    RESULT_GUARD_POSIX(s2n_stuffer_growable_alloc(&stuffer, 0));
-
-    uint32_t stuffer_size = TLS_HANDSHAKE_HEADER_LENGTH + handshake_message_length;
-    RESULT_GUARD_POSIX(s2n_stuffer_skip_write(&stuffer, stuffer_size));
-    RESULT_GUARD_POSIX(s2n_handshake_finish_header(&stuffer));
-
-    return S2N_RESULT_OK;
-}
-
 int main(int argc, char **argv)
 {
     BEGIN_TEST();
     EXPECT_SUCCESS(s2n_disable_tls13_in_test());
-
-    /* Test: s2n_handshake_finish_header */
-    {
-        /* When handshake meesage size is larger than uint16_max */
-        EXPECT_OK(s2n_test_handshake_finish_header(UINT16_MAX + 1));
-
-        /* When handshake message size is larger than uint24_max */
-        EXPECT_ERROR_WITH_ERRNO(s2n_test_handshake_finish_header(UINT24_MAX + 1), S2N_ERR_INTEGER_OVERFLOW);
-    }
 
     char dhparams_pem[S2N_MAX_TEST_PEM_SIZE];
     EXPECT_SUCCESS(s2n_read_test_pem(S2N_DEFAULT_TEST_DHPARAMS, dhparams_pem, S2N_MAX_TEST_PEM_SIZE));
